feat: commented top-level code
