+++ src/main.ts
@@ -1,12 +1,7 @@
 import { app, BrowserWindow } from 'electron';
 import path from 'path';
 import started from 'electron-squirrel-startup';
-import IPC from './shared/ipc';
-import { Worker } from 'worker_threads';
-import { bad, good } from './shared/reply';
-import { WorkerRequest } from './shared/worker/request';
-import { WorkerResponseSchema } from './shared/worker/response';
-import z from 'zod/v4';
+import { handleIPC } from './shared/ipc/handle';
 
 /** ------------------------------------------------------------------------- */
 
@@ -14,8 +9,6 @@ import z from 'zod/v4';
 if (started) {
   app.quit();
 }
-
-const { ipcMain } = IPC;
 
 async function createWindow() {
   // Create the browser window.
@@ -31,106 +24,8 @@ async function createWindow() {
     icon: './images/icon.png'
   });
 
-  // Allow all handlers.
-  ipcMain.handle.chooseDir();
-  ipcMain.handle.getPing();
-  ipcMain.handle.getSettings();
-  ipcMain.handle.setSettings();
-  ipcMain.handle.openDir();
-  ipcMain.handle.getTransformers();
-  ipcMain.handle.openOutputFile();
-  ipcMain.handle.getAllQuarters();
-  ipcMain.handle.createQuarter();
-
-  // Talking with system.
-  let worker: Maybe<Worker>;
-
-  ipcMain.handle.answerQuestion(async (_, { data }) => {
-    if (worker == null) return bad("System is not running!");
-
-    worker.postMessage({ type: "answer", ...data } as WorkerRequest);
-    return good(undefined);
-  });
-
-  ipcMain.handle.cancelProgram(async () => {
-    if (worker == null) return bad("System is not running!");
-
-    await worker.terminate();
-    worker = null;
-
-    return good(undefined);
-  });
-
-  ipcMain.handle.exitProgram(async () => {
-    if (worker == null) return bad("System is not running!");
-
-    worker.postMessage({ type: "exit" } as WorkerRequest);
-    return good(undefined);
-  });
-
-  ipcMain.handle.ignoreAll(async () => {
-    if (worker == null) return bad("System is not running!");
-    
-    worker.postMessage({ type: "ignore_all" } as WorkerRequest);
-    return good(undefined);
-  })
-
-  ipcMain.handle.runProgram(async (_, { data }) => {
-    if (worker != null) {
-      await worker.terminate();
-      worker = null;
-    }
-    
-    if (data == null) {
-      return bad("Cannot give empty settings.");
-    }
-
-    worker = new Worker(path.join(__dirname, 'worker.js'), {
-      workerData: data
-    });
-    
-    worker.on("message", message => {
-      const response_parse = WorkerResponseSchema.safeParse(message);
-      if (!response_parse.success) {
-        ipcMain.invoke.runnerUpdate(mainWindow, { type: "error", message: z.prettifyError(response_parse.error) });
-        return;
-      }
-
-      const { data } = response_parse;
-
-      switch (data.type) {
-        case "status": {
-          if (data.status.type === "error" || data.status.type === "done") {
-            worker?.terminate();
-          }
-
-          ipcMain.invoke.runnerUpdate(mainWindow, data.status);
-        } break;
-        case "question": {
-          ipcMain.invoke.runnerQuestion(mainWindow, data);
-        }
-      }
-    });
-
-    worker.on("exit", () => worker = null);
-
-    return good(undefined);
-  });
-
-  mainWindow.on("close", () => {
-    ipcMain.remove.answerQuestion();
-    ipcMain.remove.chooseDir();
-    ipcMain.remove.getPing();
-    ipcMain.remove.getSettings();
-    ipcMain.remove.getTransformers();
-    ipcMain.remove.openDir();
-    ipcMain.remove.runProgram();
-    ipcMain.remove.setSettings();
-    ipcMain.remove.openOutputFile();
-    ipcMain.remove.getAllQuarters();
-    ipcMain.remove.createQuarter();
-    ipcMain.remove.ignoreAll();
-  });
+  // Allow all handlers
+  handleIPC(mainWindow);
 
   // and load the index.html of the app.
   if (MAIN_WINDOW_VITE_DEV_SERVER_URL) {
