+++ src/preload.ts
@@ -2,6 +2,7 @@ import IPC from "./shared/ipc";
 
 /** ------------------------------------------------------------------------- */
 
+// Instantiate IPC.
 const API = IPC.exposeApiToGlobalWindow({ exposeAll: true });
 
 declare global {
